Other fixes and remove certain modules for now
Actually the way python handles certain std modules is pretty clever. The ones which don't need any kind of external dependency are written inside the source code (math for example) while others are a mixture of python code and python shared libraries (.pyd)

+++ src/tea_time.c
@@ -12,7 +12,7 @@ static TeaValue clock_time(TeaVM* vm, int count, TeaValue* args)
         return EMPTY_VAL;
     }
 
-    return NUMBER_VAL((double)clock() / CLOCKS_PER_SEC);
+    return NUMBER_VAL(((double)clock()) / ((double)CLOCKS_PER_SEC));
 }
 
 static TeaValue time_time(TeaVM* vm, int count, TeaValue* args)
